frontend/requirements.txt
Flask==2.1.1
python-dotenv
flask-restful
requests
flask-login
flask_wtf
email_validator
PyJWT